Drop embedded-gui requirement and change package layout
embedded-gui was only used for `snip-width` and `snip-height` functions, but
it links in 20Mb worth of other modules (most of framework and drracket), so
these functions were implemented directly.

Also changed package layout to avoid having a "private" folder, with a
"main.rkt" just importing the private package and re-exporting its functions.